refactor(admin) move admin api specific code to helpers
- Admin API lapis handlers have been moved to api_helpers
- `/status` handler has been moved to a dedicated module

The code being refactored will be reused in a following commit for a new
status server module.
