+++ kong/api/api_helpers.lua
@@ -1,10 +1,21 @@
 local utils = require "kong.tools.utils"
 local cjson = require "cjson"
+local pl_pretty = require "pl.pretty"
+local tablex = require "pl.tablex"
+local app_helpers = require "lapis.application"
+local arguments = require "kong.api.arguments"
+local Errors = require "kong.db.errors"
+local singletons = require "kong.singletons"
 
-local type = type
-local pairs = pairs
+local ngx      = ngx
+local sub      = string.sub
+local find     = string.find
+local type     = type
+local pairs    = pairs
+local ipairs   = ipairs
 
 local _M = {}
+local NO_ARRAY_INDEX_MARK = {}
 
 -- Parses a form value, handling multipart/data values
 -- @param `v` The value object
@@ -13,7 +24,6 @@ local function parse_value(v)
   return type(v) == "table" and v.content or v -- Handle multipart
 end
 
-local NO_ARRAY_INDEX_MARK = {}
 
 -- given a string like "x[1].y", return an array of indices like {"x", 1, "y"}
 -- the path parameter is an output-only param. the keys are added to it in order
@@ -191,6 +201,229 @@ do
     return { fields = fields }
   end
   _M.schema_to_jsonable = schema_to_jsonable
+end
+
+
+local NEEDS_BODY = tablex.readonly({ PUT = 1, POST = 2, PATCH = 3 })
+
+
+function _M.before_filter(self)
+  if not NEEDS_BODY[ngx.req.get_method()] then
+    return
+  end
+
+  local content_type = self.req.headers["content-type"]
+  if not content_type then
+    local content_length = self.req.headers["content-length"]
+    if content_length == "0" then
+      return
+    end
+
+    if not content_length then
+      local _, err = ngx.req.socket()
+      if err == "no body" then
+        return
+      end
+    end
+
+  elseif sub(content_type, 1, 16) == "application/json"                  or
+         sub(content_type, 1, 19) == "multipart/form-data"               or
+         sub(content_type, 1, 33) == "application/x-www-form-urlencoded" then
+    return
+  end
+
+  return kong.response.exit(415)
+end
+
+
+local function parse_params(fn)
+  return app_helpers.json_params(function(self, ...)
+    if NEEDS_BODY[ngx.req.get_method()] then
+      local content_type = self.req.headers["content-type"]
+      if content_type then
+        content_type = content_type:lower()
+
+        if find(content_type, "application/json", 1, true) and not self.json then
+          return kong.response.exit(400, { message = "Cannot parse JSON body" })
+
+        elseif find(content_type, "application/x-www-form-urlencode", 1, true) then
+          self.params = utils.decode_args(self.params)
+        end
+      end
+    end
+
+    self.params = _M.normalize_nested_params(self.params)
+
+    local res, err = fn(self, ...)
+
+    if err then
+      kong.log.err(err)
+      return ngx.exit(500)
+    end
+
+    if res == nil and ngx.status >= 200 then
+      return ngx.exit(0)
+    end
+
+    return res
+  end)
+end
+
+
+-- new DB
+local function new_db_on_error(self)
+  local err = self.errors[1]
+
+  if type(err) ~= "table" then
+    kong.log.err(err)
+    return kong.response.exit(500, { message = "An unexpected error occurred" })
+  end
+
+  if err.strategy then
+    err.strategy = nil
+  end
+
+  if err.code == Errors.codes.SCHEMA_VIOLATION
+  or err.code == Errors.codes.INVALID_PRIMARY_KEY
+  or err.code == Errors.codes.FOREIGN_KEY_VIOLATION
+  or err.code == Errors.codes.INVALID_OFFSET
+  or err.code == Errors.codes.FOREIGN_KEYS_UNRESOLVED
+  then
+    return kong.response.exit(400, err)
+  end
+
+  if err.code == Errors.codes.NOT_FOUND then
+    return kong.response.exit(404, err)
+  end
+
+  if err.code == Errors.codes.OPERATION_UNSUPPORTED then
+    kong.log.err(err)
+    return kong.response.exit(405, err)
+  end
+
+  if err.code == Errors.codes.PRIMARY_KEY_VIOLATION
+  or err.code == Errors.codes.UNIQUE_VIOLATION
+  then
+    return kong.response.exit(409, err)
+  end
+
+  kong.log.err(err)
+  return kong.response.exit(500, { message = "An unexpected error occurred" })
+end
+
+
+-- old DAO
+local function on_error(self)
+  local err = self.errors[1]
+
+  if type(err) ~= "table" then
+    kong.log.err(err)
+    return kong.response.exit(500, { message = "An unexpected error occurred" })
+  end
+
+  if err.name then
+    return new_db_on_error(self)
+  end
+
+  if err.db then
+    kong.log.err(err.message)
+    return kong.response.exit(500, { message = "An unexpected error occurred" })
+  end
+
+  if err.unique then
+    return kong.response.exit(409, err.tbl)
+  end
+
+  if err.foreign then
+    return kong.response.exit(404, err.tbl or { message = "Not found" })
+  end
+
+  return kong.response.exit(400, err.tbl or err.message)
+end
+
+
+local handler_helpers = {
+  yield_error = app_helpers.yield_error
+}
+
+
+function _M.attach_routes(app, routes)
+  for route_path, methods in pairs(routes) do
+    methods.on_error = methods.on_error or on_error
+
+    for method_name, method_handler in pairs(methods) do
+      local wrapped_handler = function(self)
+        return method_handler(self, {}, handler_helpers)
+      end
+
+      methods[method_name] = parse_params(wrapped_handler)
+    end
+
+    app:match(route_path, route_path, app_helpers.respond_to(methods))
+  end
+end
+
+
+function _M.attach_new_db_routes(app, routes)
+  for route_path, definition in pairs(routes) do
+    local schema  = definition.schema
+    local methods = definition.methods
+
+    methods.on_error = methods.on_error or new_db_on_error
+
+    for method_name, method_handler in pairs(methods) do
+      local wrapped_handler = function(self)
+        self.args = arguments.load({
+          schema  = schema,
+          request = self.req,
+        })
+
+        return method_handler(self, singletons.db, handler_helpers)
+      end
+
+      methods[method_name] = parse_params(wrapped_handler)
+    end
+
+    app:match(route_path, route_path, app_helpers.respond_to(methods))
+  end
+end
+
+
+function _M.default_route(self)
+  local path = self.req.parsed_url.path:match("^(.*)/$")
+
+  if path and self.app.router:resolve(path, self) then
+    return
+
+  elseif self.app.router:resolve(self.req.parsed_url.path .. "/", self) then
+    return
+  end
+
+  return self.app.handle_404(self)
+end
+
+
+
+function _M.handle_404(self)
+  return kong.response.exit(404, { message = "Not found" })
+end
+
+
+function _M.handle_error(self, err, trace)
+  if err then
+    if type(err) ~= "string" then
+      err = pl_pretty.write(err)
+    end
+    if find(err, "don't know how to respond to", nil, true) then
+      return kong.response.exit(405, { message = "Method not allowed" })
+    end
+  end
+
+  ngx.log(ngx.ERR, err, "\n", trace)
+
+  -- We just logged the error so no need to give it to responses and log it
+  -- twice
+  return kong.response.exit(500, { message = "An unexpected error occurred" })
 end
